requirements.txt
EasyProcess
PyHamcrest
PyVirtualDisplay
SQLAlchemy
argparse
execnet
fuzzywuzzy
jinja2
lettuce
pika
psycopg2
python-ldap
requests>=2.0
selenium
sure
wsgiref
pexpect
git+git://github.com/xivo-pbx/pylinphonelib.git@master
git+git://github.com/xivo-pbx/xivo-ws.git
git+https://github.com/xivo-pbx/xivo-dao.git
git+https://github.com/xivo-pbx/xivo-provisioning.git
git+https://github.com/xivo-pbx/xivo-bus.git
git+https://github.com/xivo-pbx/xivo-lib-python.git
